View based requests
Created filter which allows the view to 'getValue()' from the model, rather than relying on the controller.
Also created a filter for generating a range of pagination indexes, based on a current page, page size, and max.
This should make FalcorJS look as good in Angular v1 as it does in v2.